pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Voice-enabled password reset agent using OpenAI"
requires-python = ">=3.9"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "numpy>=2.0.2",
    "openai>=1.12.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.6.3",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.2",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"] 
